Create fundamentals.py
These practice problems have little to no explanation with the intention of making any beginner look deeper in to the problem an utilize online resources to find the answer. Or, to see if they can find the patterns that exist in writing these lines of code.

Once I introduce user made functions ill provide explanations for what is going on in the logic. 

If you are a beginner, great job for practicing. If you are more experienced I would love some feed back on anything you think would like to add!